requirements-train.txt
ultralytics
pycocotools
opencv-python
tqdm
pyyaml
Pillow
